iter57 cluster-072: Studio catalog storage Protobuf serializers(#955)
issue #955 direct implement(medium severity,no design):
- Studio catalog 存储改用 Protobuf 序列化(满足 CLAUDE '统一 Protobuf' 强制)
- 新 StudioCatalogProtobufStorageSerializer + 测试
- 删除原 JSON serializer(违反 '禁止 JSON 用于持久态')

closes #955
local PASS:architecture + test_stability + build + StudioCatalogProtobufStorageSerializerTests

⟦AI:AUTO-LOOP⟧
