+++ src/Aevatar.Studio.Infrastructure/Storage/ConnectorCatalogStorageSerializer.cs
@@ -1,4 +1,3 @@
-using System.Text.Json;
 using Aevatar.GAgents.ConnectorCatalog;
 using Aevatar.Studio.Application.Studio.Abstractions;
 using Google.Protobuf;
@@ -8,30 +7,18 @@ namespace Aevatar.Studio.Infrastructure.Storage;
 
 internal static class ConnectorCatalogStorageSerializer
 {
-    // Refactor (iter22/cluster-001-studio-json-internal-catalog-storage):
-    //   Old pattern: Studio connector catalog and draft facts were durable JSON documents.
-    //   New principle: Durable storage payloads are protobuf facts; JSON is only import fallback.
     public static async Task<IReadOnlyList<StoredConnectorDefinition>> ReadCatalogAsync(
         Stream stream,
         CancellationToken cancellationToken)
     {
         var payload = await ReadPayloadAsync(stream, cancellationToken);
-        if (!IsJsonPayload(payload))
-        {
-            var state = ConnectorCatalogState.Parser.ParseFrom(payload);
-            return state.Connectors
-                .Select(ToStoredConnectorDefinition)
-                .ToList()
-                .AsReadOnly();
-        }
-
-        using var document = JsonDocument.Parse(payload);
-        return ParseConnectors(document.RootElement);
+        var state = ConnectorCatalogState.Parser.ParseFrom(payload);
+        return state.Connectors
+            .Select(ToStoredConnectorDefinition)
+            .ToList()
+            .AsReadOnly();
     }
 
-    // Refactor (iter22/cluster-001-studio-json-internal-catalog-storage):
-    //   Old pattern: Studio connector catalog writes emitted durable JSON.
-    //   New principle: Catalog writes emit the protobuf catalog state fact.
     public static async Task WriteCatalogAsync(
         Stream stream,
         IReadOnlyList<StoredConnectorDefinition> connectors,
@@ -42,39 +29,18 @@ internal static class ConnectorCatalogStorageSerializer
         await stream.WriteAsync(payload.ToByteArray(), cancellationToken);
     }
 
-    // Refactor (iter22/cluster-001-studio-json-internal-catalog-storage):
-    //   Old pattern: Studio connector draft reads treated JSON as the durable format.
-    //   New principle: Draft reads prefer protobuf and keep JSON only as a bounded legacy import fallback.
     public static async Task<ParsedConnectorDraft> ReadDraftAsync(
         Stream stream,
         DateTimeOffset fallbackUpdatedAtUtc,
         CancellationToken cancellationToken)
     {
         var payload = await ReadPayloadAsync(stream, cancellationToken);
-        if (!IsJsonPayload(payload))
-        {
-            var draftEntry = ConnectorDraftEntry.Parser.ParseFrom(payload);
-            var protobufUpdatedAtUtc = draftEntry.UpdatedAtUtc?.ToDateTimeOffset() ?? fallbackUpdatedAtUtc;
-            var protobufDraft = draftEntry.Draft is not null ? ToStoredConnectorDefinition(draftEntry.Draft) : null;
-            return new ParsedConnectorDraft(protobufUpdatedAtUtc, protobufDraft);
-        }
-
-        using var document = JsonDocument.Parse(payload);
-        var root = document.RootElement;
-        var updatedAtUtc = TryGetPropertyIgnoreCase(root, "updatedAtUtc", out var updatedAtNode) &&
-                           updatedAtNode.ValueKind == JsonValueKind.String &&
-                           DateTimeOffset.TryParse(updatedAtNode.GetString(), out var parsedUpdatedAt)
-            ? parsedUpdatedAt
-            : fallbackUpdatedAtUtc;
-
-        var draftNode = TryGetPropertyIgnoreCase(root, "connector", out var connectorNode) ? connectorNode : root;
-        var draft = draftNode.ValueKind == JsonValueKind.Object ? ParseConnector(draftNode, null) : null;
-        return new ParsedConnectorDraft(updatedAtUtc, draft);
+        var draftEntry = ConnectorDraftEntry.Parser.ParseFrom(payload);
+        var protobufUpdatedAtUtc = draftEntry.UpdatedAtUtc?.ToDateTimeOffset() ?? fallbackUpdatedAtUtc;
+        var protobufDraft = draftEntry.Draft is not null ? ToStoredConnectorDefinition(draftEntry.Draft) : null;
+        return new ParsedConnectorDraft(protobufUpdatedAtUtc, protobufDraft);
     }
 
-    // Refactor (iter22/cluster-001-studio-json-internal-catalog-storage):
-    //   Old pattern: Studio connector draft writes emitted durable JSON.
-    //   New principle: Draft writes emit the protobuf draft fact.
     public static async Task WriteDraftAsync(
         Stream stream,
         StoredConnectorDefinition? draft,
@@ -99,21 +65,6 @@ internal static class ConnectorCatalogStorageSerializer
         using var buffer = new MemoryStream();
         await stream.CopyToAsync(buffer, cancellationToken);
         return buffer.ToArray();
-    }
-
-    private static bool IsJsonPayload(ReadOnlySpan<byte> payload)
-    {
-        foreach (var value in payload)
-        {
-            if (value is (byte)' ' or (byte)'\t' or (byte)'\r' or (byte)'\n')
-            {
-                continue;
-            }
-
-            return value is (byte)'{' or (byte)'[';
-        }
-
-        return false;
     }
 
     private static StoredConnectorDefinition ToStoredConnectorDefinition(ConnectorDefinitionEntry entry) =>
@@ -231,146 +182,6 @@ internal static class ConnectorCatalogStorageSerializer
         }
     }
 
-    private static IReadOnlyList<StoredConnectorDefinition> ParseConnectors(JsonElement root)
-    {
-        var connectorsNode = TryGetPropertyIgnoreCase(root, "connectors", out var configuredNode)
-            ? configuredNode
-            : root;
-
-        var results = new List<StoredConnectorDefinition>();
-        if (connectorsNode.ValueKind == JsonValueKind.Array)
-        {
-            foreach (var item in connectorsNode.EnumerateArray())
-            {
-                if (item.ValueKind != JsonValueKind.Object)
-                {
-                    continue;
-                }
-
-                var connector = ParseConnector(item, null);
-                if (connector is not null)
-                {
-                    results.Add(connector);
-                }
-            }
-
-            return results;
-        }
-
-        if (connectorsNode.ValueKind != JsonValueKind.Object)
-        {
-            return [];
-        }
-
-        if (TryGetPropertyIgnoreCase(connectorsNode, "definitions", out var definitionsNode) &&
-            definitionsNode.ValueKind == JsonValueKind.Array)
-        {
-            foreach (var item in definitionsNode.EnumerateArray())
-            {
-                if (item.ValueKind != JsonValueKind.Object)
-                {
-                    continue;
-                }
-
-                var connector = ParseConnector(item, null);
-                if (connector is not null)
-                {
-                    results.Add(connector);
-                }
-            }
-
-            return results;
-        }
-
-        foreach (var property in connectorsNode.EnumerateObject())
-        {
-            if (property.Value.ValueKind != JsonValueKind.Object)
-            {
-                continue;
-            }
-
-            var connector = ParseConnector(property.Value, property.Name);
-            if (connector is not null)
-            {
-                results.Add(connector);
-            }
-        }
-
-        return results;
-    }
-
-    private static StoredConnectorDefinition? ParseConnector(JsonElement connectorNode, string? fallbackName)
-    {
-        var name = ReadString(connectorNode, "name");
-        if (string.IsNullOrWhiteSpace(name))
-        {
-            name = fallbackName ?? string.Empty;
-        }
-
-        var type = ReadString(connectorNode, "type");
-        if (string.IsNullOrWhiteSpace(name) || string.IsNullOrWhiteSpace(type))
-        {
-            return null;
-        }
-
-        return new StoredConnectorDefinition(
-            Name: name,
-            Type: type,
-            Enabled: ReadBool(connectorNode, "enabled", true),
-            TimeoutMs: Math.Clamp(ReadInt(connectorNode, "timeoutMs", 30_000), 100, 300_000),
-            Retry: Math.Clamp(ReadInt(connectorNode, "retry", 0), 0, 5),
-            Http: TryGetPropertyIgnoreCase(connectorNode, "http", out var httpNode) ? ParseHttpConfig(httpNode) : EmptyHttpConfig(),
-            Cli: TryGetPropertyIgnoreCase(connectorNode, "cli", out var cliNode) ? ParseCliConfig(cliNode) : EmptyCliConfig(),
-            Mcp: TryGetPropertyIgnoreCase(connectorNode, "mcp", out var mcpNode) ? ParseMcpConfig(mcpNode) : EmptyMcpConfig());
-    }
-
-    private static StoredHttpConnectorConfig ParseHttpConfig(JsonElement node) =>
-        node.ValueKind != JsonValueKind.Object
-            ? EmptyHttpConfig()
-            : new StoredHttpConnectorConfig(
-                BaseUrl: ReadString(node, "baseUrl"),
-                AllowedMethods: ReadStringArray(node, "allowedMethods"),
-                AllowedPaths: ReadStringArray(node, "allowedPaths"),
-                AllowedInputKeys: ReadStringArray(node, "allowedInputKeys"),
-                DefaultHeaders: ReadStringMap(node, "defaultHeaders"),
-                Auth: TryGetPropertyIgnoreCase(node, "auth", out var authNode) ? ParseAuthConfig(authNode) : EmptyAuthConfig());
-
-    private static StoredCliConnectorConfig ParseCliConfig(JsonElement node) =>
-        node.ValueKind != JsonValueKind.Object
-            ? EmptyCliConfig()
-            : new StoredCliConnectorConfig(
-                Command: ReadString(node, "command"),
-                FixedArguments: ReadStringArray(node, "fixedArguments"),
-                AllowedOperations: ReadStringArray(node, "allowedOperations"),
-                AllowedInputKeys: ReadStringArray(node, "allowedInputKeys"),
-                WorkingDirectory: ReadString(node, "workingDirectory"),
-                Environment: ReadStringMap(node, "environment"));
-
-    private static StoredMcpConnectorConfig ParseMcpConfig(JsonElement node) =>
-        node.ValueKind != JsonValueKind.Object
-            ? EmptyMcpConfig()
-            : new StoredMcpConnectorConfig(
-                ServerName: ReadString(node, "serverName"),
-                Command: ReadString(node, "command"),
-                Url: ReadString(node, "url"),
-                Arguments: ReadStringArray(node, "arguments"),
-                Environment: ReadStringMap(node, "environment"),
-                AdditionalHeaders: ReadStringMap(node, "additionalHeaders"),
-                Auth: TryGetPropertyIgnoreCase(node, "auth", out var authNode) ? ParseAuthConfig(authNode) : EmptyAuthConfig(),
-                DefaultTool: ReadString(node, "defaultTool"),
-                AllowedTools: ReadStringArray(node, "allowedTools"),
-                AllowedInputKeys: ReadStringArray(node, "allowedInputKeys"));
-
-    private static StoredConnectorAuthConfig ParseAuthConfig(JsonElement node) =>
-        node.ValueKind != JsonValueKind.Object
-            ? EmptyAuthConfig()
-            : new StoredConnectorAuthConfig(
-                Type: ReadString(node, "type"),
-                TokenUrl: ReadString(node, "tokenUrl"),
-                ClientId: ReadString(node, "clientId"),
-                ClientSecret: ReadString(node, "clientSecret"),
-                Scope: ReadString(node, "scope"));
-
     private static StoredHttpConnectorConfig EmptyHttpConfig() =>
         new(string.Empty, [], [], [], new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase), EmptyAuthConfig());
 
@@ -392,113 +203,4 @@ internal static class ConnectorCatalogStorageSerializer
 
     private static StoredConnectorAuthConfig EmptyAuthConfig() =>
         new(string.Empty, string.Empty, string.Empty, string.Empty, string.Empty);
-
-    private static bool TryGetPropertyIgnoreCase(JsonElement element, string propertyName, out JsonElement value)
-    {
-        if (element.ValueKind != JsonValueKind.Object)
-        {
-            value = default;
-            return false;
-        }
-
-        foreach (var property in element.EnumerateObject())
-        {
-            if (string.Equals(property.Name, propertyName, StringComparison.OrdinalIgnoreCase))
-            {
-                value = property.Value;
-                return true;
-            }
-        }
-
-        value = default;
-        return false;
-    }
-
-    private static string ReadString(JsonElement element, params string[] propertyNames)
-    {
-        foreach (var propertyName in propertyNames)
-        {
-            if (TryGetPropertyIgnoreCase(element, propertyName, out var value) &&
-                value.ValueKind == JsonValueKind.String)
-            {
-                return value.GetString() ?? string.Empty;
-            }
-        }
-
-        return string.Empty;
-    }
-
-    private static bool ReadBool(JsonElement element, string propertyName, bool fallback)
-    {
-        if (!TryGetPropertyIgnoreCase(element, propertyName, out var value))
-        {
-            return fallback;
-        }
-
-        if (value.ValueKind == JsonValueKind.True)
-        {
-            return true;
-        }
-
-        if (value.ValueKind == JsonValueKind.False)
-        {
-            return false;
-        }
-
-        return value.ValueKind == JsonValueKind.String && bool.TryParse(value.GetString(), out var parsed)
-            ? parsed
-            : fallback;
-    }
-
-    private static int ReadInt(JsonElement element, string propertyName, int fallback)
-    {
-        if (!TryGetPropertyIgnoreCase(element, propertyName, out var value))
-        {
-            return fallback;
-        }
-
-        if (value.ValueKind == JsonValueKind.Number && value.TryGetInt32(out var numberValue))
-        {
-            return numberValue;
-        }
-
-        return value.ValueKind == JsonValueKind.String && int.TryParse(value.GetString(), out numberValue)
-            ? numberValue
-            : fallback;
-    }
-
-    private static IReadOnlyList<string> ReadStringArray(JsonElement element, string propertyName)
-    {
-        if (!TryGetPropertyIgnoreCase(element, propertyName, out var value) ||
-            value.ValueKind != JsonValueKind.Array)
-        {
-            return [];
-        }
-
-        return value.EnumerateArray()
-            .Where(item => item.ValueKind == JsonValueKind.String)
-            .Select(item => item.GetString() ?? string.Empty)
-            .Where(item => !string.IsNullOrWhiteSpace(item))
-            .ToList();
-    }
-
-    private static IReadOnlyDictionary<string, string> ReadStringMap(JsonElement element, string propertyName)
-    {
-        if (!TryGetPropertyIgnoreCase(element, propertyName, out var value) ||
-            value.ValueKind != JsonValueKind.Object)
-        {
-            return new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase);
-        }
-
-        var result = new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase);
-        foreach (var property in value.EnumerateObject())
-        {
-            result[property.Name] = property.Value.ValueKind == JsonValueKind.String
-                ? property.Value.GetString() ?? string.Empty
-                : property.Value.ToString();
-        }
-
-        return result;
-    }
-
 }
